refactor(*): ♻️ central source of truth for valid issue prefixes
