+++ src/prompts/lang.ts
@@ -1,6 +1,7 @@
 import { styleText } from "node:util";
 
-import type { IssuesPrefixes, PromptsLang } from "@/interfaces";
+import type { IssuesPrefixes } from "@/defaults/config";
+import type { PromptsLang } from "@/interfaces";
 
 const breaking = styleText("red", "BREAKING CHANGE:");
